JUC学习：Lock使用
