--- study-juc/src/main/java/org/ohx/studyjuc/Executors/MyThreadFactory.java
@@ -1,14 +1,0 @@
-package org.ohx.studyjuc.Executors;
-
-import java.util.concurrent.ThreadFactory;
-
-public class MyThreadFactory implements ThreadFactory {
-
-    @Override
-    public Thread newThread(Runnable r) {
-        Thread thread = new Thread(r);
-        thread.setName("定制线程名" + Math.random());
-        return thread;
-    }
-
-}
