--- study-juc/src/main/java/org/ohx/studyjuc/Executors/Run6.java
@@ -1,38 +1,0 @@
-package org.ohx.studyjuc.Executors;
-
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-
-/**
- * Executors.newSingleThreadExecutor创建单一线程池
- */
-public class Run6 {
-
-    public static void main(String[] args) {
-        ExecutorService executorService = Executors.newSingleThreadExecutor();
-        for (int i = 0; i < 3; i++) {
-            executorService.execute(new Run5.MyRunnable(i+1+""));
-        }
-    }
-
-    static class MyRunnable implements Runnable {
-
-        private String name;
-
-        public MyRunnable(String name) {
-            super();
-            this.name = name;
-        }
-
-        @Override
-        public void run() {
-            try {
-                System.out.println(Thread.currentThread().getName() + " name=" + name + " bgein " + System.currentTimeMillis());
-                Thread.sleep(2000);
-                System.out.println(Thread.currentThread().getName() + " name=" + name + " end " + System.currentTimeMillis());
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-}
