+++ study-juc/src/main/java/org/ohx/studyjuc/lock/reentrantlock/useconditionwaitnotify/Service.java
@@ -1,0 +1,36 @@
+package org.ohx.studyjuc.lock.reentrantlock.useconditionwaitnotify;
+
+import java.util.concurrent.locks.Condition;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
+
+/**
+ * @author mudkip
+ * @date 2022/6/27
+ */
+public class Service {
+    private Lock lock = new ReentrantLock();
+    private Condition condition = lock.newCondition();
+
+    public void await() {
+        try {
+            lock.lock();
+            System.out.println("await time " + System.currentTimeMillis());
+            condition.await();
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    public void signal() {
+        try {
+            lock.lock();
+            System.out.println("signal time " + System.currentTimeMillis());
+            condition.signal();
+        } finally {
+            lock.unlock();
+        }
+    }
+}
