+++ study-juc/src/main/java/org/ohx/studyjuc/threadpool/executors/MyThreadFactory.java
@@ -1,0 +1,14 @@
+package org.ohx.studyjuc.threadpool.executors;
+
+import java.util.concurrent.ThreadFactory;
+
+public class MyThreadFactory implements ThreadFactory {
+
+    @Override
+    public Thread newThread(Runnable r) {
+        Thread thread = new Thread(r);
+        thread.setName("定制线程名" + Math.random());
+        return thread;
+    }
+
+}
